Finished initial DAO classes. Got started on the Controllers.

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 
 public interface AccountDao {
 
-    BigDecimal viewBalance(int userId);
+    BigDecimal viewBalance(Long userId);
 
     void update(Account accountToUpdate);
