+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/UserDao.java
@@ -10,8 +10,9 @@ public interface UserDao {
 
     User findByUsername(String username);
 
-    int findIdByUsername(String username);
+    Long findIdByUsername(String username);
 
+    // As a user of the system, I need to be able to register myself with a username and password.
     boolean create(String username, String password);
 
     User update(User userToUpdate);
